# Faster implementation of `InputBuffer`