--- a/Cargo.toml
+++ b/Cargo.toml
@@ -27,7 +27,6 @@ byteorder = "1.3.2"
 bytes = "1.0"
 http = "0.2"
 httparse = "1.3.4"
-input_buffer = "0.4.0"
 log = "0.4.8"
 rand = "0.8.0"
 sha-1 = "0.9"
@@ -53,5 +52,6 @@ optional = true
 version = "0.5.0"
 
 [dev-dependencies]
+input_buffer = "0.5.0"
 env_logger = "0.8.1"
 net2 = "0.2.33"