--- a/Cargo.toml
+++ b/Cargo.toml
@@ -52,6 +52,12 @@ optional = true
 version = "0.5.0"
 
 [dev-dependencies]
-input_buffer = "0.5.0"
+criterion = "0.3.4"
 env_logger = "0.8.1"
+input_buffer = "0.5.0"
 net2 = "0.2.33"
+rand = "0.8.4"
+
+[[bench]]
+name = "buffer"
+harness = false