--- a/src/client.rs
+++ b/src/client.rs
@@ -72,7 +72,8 @@ mod encryption {
             Mode::Tls => {
                 let config = {
                     let mut config = ClientConfig::new();
-                    config.root_store = rustls_native_certs::load_native_certs().map_err(|(_, err)| err)?;
+                    config.root_store =
+                        rustls_native_certs::load_native_certs().map_err(|(_, err)| err)?;
 
                     Arc::new(config)
                 };