Fixes https://github.com/snapview/tungstenite-rs/issues/210.

Please [check the discussion and decision](https://github.com/snapview/input_buffer/issues/6#issuecomment-870548303) if you don't know the context as well as the [latest PR in `input_buffer`](https://github.com/snapview/input_buffer/pull/9).

It seems like the new simplified implementation has only significant difference on x86_64 architecture while the changes on Armv8.4-A A64 (Apple M1) are not that significant as both versions perform really well. On ARM the new version even behaves a bit slower (but not significantly slower), however the difference on Intel (x86_64) architecture is quite significant.

```
===== M1 Mac (Armv8.4-A A64) =====

buffers/InputBuffer     time:   [543.66 us 549.32 us 555.54 us]
                        thrpt:  [7.0315 GiB/s 7.1110 GiB/s 7.1851 GiB/s]

buffers/ReadBuffer      time:   [552.16 us 565.83 us 581.26 us]
                        thrpt:  [6.7203 GiB/s 6.9036 GiB/s 7.0745 GiB/s]


===== Intel (x86_64) =====

buffers/InputBuffer     time:   [4.7392 ms 4.7601 ms 4.7817 ms]
                        thrpt:  [836.52 MiB/s 840.31 MiB/s 844.03 MiB/s]

buffers/ReadBuffer      time:   [726.86 us 730.72 us 734.81 us]
                        thrpt:  [5.3160 GiB/s 5.3457 GiB/s 5.3741 GiB/s]
```

So it seems like migrating to the newer `ReadBuffer` makes sense as the 'average performance gain' is quite significant (and I believe most users are still using x86_64).

⚠️ Notes:
1. The PR also removes a couple of errors that could have been triggered by `tungstenite-rs`. These were removed as they could only occur if the incoming message does not fit into the hardcoded `u32::max_value()`. Currently we have [`max_frame_size` and `max_message_size`](https://docs.rs/tungstenite/0.13.0/tungstenite/protocol/struct.WebSocketConfig.html), but we don't use them to cap the size of the input/read buffer (instead we compare the values once we read something). This in theory is not a problem when using `tungstenite-rs` with non-blocking sockets (like in case of `tokio-tungstenite`), though it could pose the issue for those ones who use blocking sockets. In other words, the previous reading with `u32::max_value()` did not really add any sort of "additional protection", so I removed them all together to not complicate the `ReadBuffer` implementation.
2. There is no automation to run `criterion` benchmarks with our CI. That's fine as `criterion` is not meant to be used like this and may give many false positives and false negatives when used like this. In order to automate benchmarking with CI we have to use a tool like [lai](https://github.com/bheisler/iai) that was written to be used for such purposes (it works differently than `criterion`), but I think it's better to make it as part of a separate PR.


P.S.: Thanks @qiujiangkun for bringing this topic and performance testing of `tungstenite-rs` 👍