@agalakhov 2021-07-05
Is there any size limit now? Is it possible to DoS by memory overfull?