@application-developer-DA 2021-07-06
> Is there any size limit now? Is it possible to DoS by memory overfull?

Do you mean those points that I mentioned in (1)? If so, then I think there is no problem with it since the previous limit that we had on the `InputBuffer` was a hardcoded value of `usize::max_value()` which I believe could not even be triggered in any normal circumstances because of [`max_frame_size` and `max_message_size`](https://docs.rs/tungstenite/0.13.0/tungstenite/protocol/struct.WebSocketConfig.html) that were checked (and still checked) during the message assembly. Since we read chunks, the only way to trigger the error from the `InputBuffer` was user setting either of `max_frame_size` or `max_message_size` in the `WebSocketConfig` to the `Some(usize::MAX)`. Another case when the limit could hit was the initial handshake. I don't think that the new implementation is worse than **previous** implementation in those particular cases since even if we get to the `usize::MAX`, the size of the underlying `Vec` will be at its maximum value (`len` is `usize`), pushing to such `Vec` would result in panic. But the **old implementation** did not handle it better since it tried to return an error by summing up and checking the `usize` values without taking the overflow into an account. Adding `usize` with some `reserve` would lead to the panic in the old implementation as well.

In other words, the new drop-in replacement is probably not perfect (since there is a way to get the panic), but it is at least not worse than the previous implementation. Please correct me if I'm wrong or missed something.

However while replying to your comment and checking the unit tests (trying to extend them) I noticed another thing: the new implementation had an issue that it did not clean up the data that is read by the cursor (which normally is not required anymore), so I pushed a new commit that implements a behavior similar to the previous `remove_garbage` stage in the `InputBuffer` and extended the unit test to account for that.

With that step implemented the new `ReadBuffer` seems to still be faster.

```
buffers/InputBuffer     time:   [3.1572 ms 3.1728 ms 3.1904 ms]                                 
                        thrpt:  [1.2244 GiB/s 1.2312 GiB/s 1.2373 GiB/s]

buffers/ReadBuffer      time:   [1.2585 ms 1.2659 ms 1.2737 ms]                                
                        thrpt:  [3.0668 GiB/s 3.0857 GiB/s 3.1038 GiB/s]
```